Add "Refresh Image" to RNTester Image.fadeDuration example (#44414)
Summary:

# Changelog:
[Internal] - 

I noticed that "Image/Fade Duration" test in RNTester is practically useless, as at the moment one scrolls to the test, the fading is most probably had already ended.

This adds a "button" to refresh the image and be able to see the fading in again and again, if desired.

Differential Revision: D56978930