Suppress LogBox for low-severity errors and warnings if Fusebox console is available
Summary:
Changelog: [Internal]

Showing warnings and low-severity errors in LogBox is noisy, confusing for web developers, and not the best use of screen real estate on mobile platforms. Since the Fusebox console offers a superior experience, as of this diff we'll suppress warnings and low-severity errors in LogBox if we detect that Fusebox is available.

*The first time* a warning or error is suppressed (i.e. at most once per app launch), we'll show a "log" level notification pointing the user towards Fusebox. When the notification is clicked, we call the `DevSettings.openDebugger` method and dismiss it.

The wording of the notification ("Open debugger to view logs") is intentional:

1. It's short enough to fit on small screens in its entirety.
2. It doesn't actually say "*click here* to open the debugger". This is for the best because `DevSettings.openDebugger` is a best-effort method that might fail, and in the current implementation there's no reliable feedback to the user about the success/failure of the launch.

Differential Revision: D57681446