Remove dependency on BridgeDevSupportManager
DevLoadingModule was using BridgeDevSupportManager to initialize DevLoadingViewManager. This meant a lot of devsupport classes were included in the relase APK.
This commit refactors DevLoadingModule to use only the DevSupportManager interface to mitigate the above problem.
DevSupportManagerBase always creates an instance of DevLoadingViewManager, which is why it need not be initialized again in DevLoadingModule.
Removing old related unused methods from BridgeDevSupportManager