Integrate perfetto with UserPerformance API (#44702)
Summary:

Based on bgirard initial changes in D53478653, this creates an initial integration of the Perfetto SDK with the User Timing API, allowing performance information to be logged from JS to Perfetto traces.

The logic in `initializePerfetto` may need to be to other targets (eg reactperflogger) once we want to make this usable in other components, but keeping it scoped to User Timing for now.

Changelog: [Internal]

Differential Revision: D57881823