Reland: Simplify ReactViewGroup clip to border
Summary:
## Reland

The original change had getBorderBoxPath() and getBorderBoxPath() returning non-nullable path, which is incorrect. The fix here is to only calculate and return path when there are rounded corners set, and otherwise force caller to handle rect  when path is null. Code in ReactViewGroup was already handling this, but code in BoxShadowDrawable was not.

The crash seemed to happen on Android API 24 emulators, where `onBoundsChanged` is not called, so `mNeedUpdatePathForBorderRadius` is never set unless some other props are set. But, we really shouldn't force path creation unless needed.

## Original

We can remove most of the code for clipping children to border radius, and recalculating paths, in ReactViewGroup, and rely on the padding box path/rect already set.

I will move this to something more generic up the stack so other native components can reuse this logic.

Changelog: [Internal]

Differential Revision: D57934907