Update runtime shadow node references when cloning within react native
Summary:
Any shadow node cloning happening outside the execution of the UIManagerBinding cloneNode call should guarantee that the references held to the shadow node are updated to the latest revision.

This guarantees that the React renderer's current fiber tree holds references to the ShadowNode instances that acually were layed out and committed for rendering on the native side. Maintaining these references up to date on the JS current fiber tree allows to maximize layout caches on subsequent commits.

Differential Revision: https://www.internalfb.com/diff/D57860867?entry_point=27