fix: Helloworld assumes community cli is present
Use the hard-coded config instead to avoid depending on the community
cli to generate a config. This fixes both iOS and Android builds.

Minor fix for the gradle-plugin to ignore a generated react-native
config if it's empty. This can happen if the command fails to generate
output. This occurs if the command fails.

Changelog: [Internal]